requirements.txt
fuzzywuzzy==0.18.0
packaging==23.2
pillow==10.2.0
pytesseract==0.3.10
Wand==0.6.13
